database finally works

+++ apps/backend/src/extractors/auth.rs
@@ -7,7 +7,7 @@ use crate::{constants::auth::AUTH_COOKIE_NAME, state::AppState};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct AuthUser {
-    pub player_id: i64,
+    pub player_id: i32,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
